Modified pneumatics code.
I changed the code so we can plug in both the lift and the gripper this
week.  I also changed where the pressure sensor will plug in; it's old
position would interfere with the newly added ultrasonic sensor.  This
is noted in a comment.

I successfully built the project.  There were no errors, but 3 warnings.
(Sage, can you take a look?)